use a second cli and a separate endpoint for the mos experiment, instead of subcommands, so that we do not break the commands deployed in production now for the standard experiments
