+++ tests/cams2_83/test_cams2_83_cli.py
@@ -32,8 +32,10 @@ def test_clearcache(
         assert var_list is None
         assert update_interface is True
 
-    monkeypatch.setattr("pyaerocom.scripts.cams2_83.cli.ExperimentProcessor.run", do_not_run)
-    options = f"main forecast week 2024-03-16 2024-03-23 --model-path {tmp_path} --obs-path {tmp_path} --data-path {tmp_path} --coldata-path {tmp_path} --name 'Test'"
+    monkeypatch.setattr(
+        "pyaerocom.scripts.cams2_83.evaluation.ExperimentProcessor.run", do_not_run
+    )
+    options = f"forecast week 2024-03-16 2024-03-23 --model-path {tmp_path} --obs-path {tmp_path} --data-path {tmp_path} --coldata-path {tmp_path} --name 'Test'"
     result = runner.invoke(app, options.split())
     assert "Running Statistics" in caplog.text
     assert result.exit_code == 0
@@ -63,8 +65,8 @@ def test_not_cleared_cache(
         assert analysis is False
         assert update_interface is True
 
-    monkeypatch.setattr("pyaerocom.scripts.cams2_83.cli.CAMS2_83_Processer.run", do_not_run)
-    options = f"main forecast long 2024-03-16 2024-03-23 --model-path {tmp_path} --obs-path {tmp_path} --data-path {tmp_path} --coldata-path {tmp_path} --name 'Test' --medianscores"
+    monkeypatch.setattr("pyaerocom.scripts.cams2_83.evaluation.CAMS2_83_Processer.run", do_not_run)
+    options = f"forecast long 2024-03-16 2024-03-23 --model-path {tmp_path} --obs-path {tmp_path} --data-path {tmp_path} --coldata-path {tmp_path} --name 'Test' --medianscores"
     result = runner.invoke(app, options.split())
     assert "Running CAMS2_83 Specific Statistics, cache is not cleared" in caplog.text
     assert result.exit_code == 0
@@ -79,7 +81,7 @@ def test_eval_dummy(
 ):
     assert list(fake_cache_path.glob("*.pkl"))
 
-    options = f"main forecast day 2024-03-16 2024-03-16 --model-path {tmp_path} --obs-path {tmp_path} --data-path {tmp_path} --coldata-path {tmp_path} --name 'Test'"
+    options = f"forecast day 2024-03-16 2024-03-16 --model-path {tmp_path} --obs-path {tmp_path} --data-path {tmp_path} --coldata-path {tmp_path} --name 'Test'"
     result = runner.invoke(app, options.split())
     assert result.exit_code == 0
     assert "Failed to read model variable" in caplog.text
@@ -92,21 +94,8 @@ def test_eval_medianscores_dummy(
 ):
     assert list(fake_cache_path.glob("*.pkl"))
 
-    options = f"main analysis long 2023-03-01 2024-02-28 --model-path {tmp_path} --obs-path {tmp_path} --data-path {tmp_path} --coldata-path {tmp_path} --name 'Test' --medianscores"
+    options = f"analysis long 2023-03-01 2024-02-28 --model-path {tmp_path} --obs-path {tmp_path} --data-path {tmp_path} --coldata-path {tmp_path} --name 'Test' --medianscores"
     result = runner.invoke(app, options.split())
     assert result.exit_code == 0
     assert "Running CAMS2_83 Specific Statistics, cache is not cleared" in caplog.text
     assert "Failed to read model variable" in caplog.text
-
-
-def test_eval_mos_dummy(
-    fake_cache_path: Path,
-    tmp_path: Path,
-    caplog,
-):
-    assert list(fake_cache_path.glob("*.pkl"))
-
-    options = f"mos season 2024-03-01 2024-05-12 --data-path {tmp_path} --coldata-path {tmp_path} --name 'Test'"
-    result = runner.invoke(app, options.split())
-    assert result.exit_code == 0
-    assert "no output available" in caplog.text
